+++ packages/caravella_core/lib/model/expense_location.dart
@@ -4,7 +4,32 @@ class ExpenseLocation {
   final String? address;
   final String? name;
 
-  ExpenseLocation({this.latitude, this.longitude, this.address, this.name});
+  // Detailed geocoding information
+  final String? street;
+  final String? streetNumber;
+  final String? locality;
+  final String? subLocality;
+  final String? administrativeArea;
+  final String? subAdministrativeArea;
+  final String? postalCode;
+  final String? country;
+  final String? isoCountryCode;
+
+  ExpenseLocation({
+    this.latitude,
+    this.longitude,
+    this.address,
+    this.name,
+    this.street,
+    this.streetNumber,
+    this.locality,
+    this.subLocality,
+    this.administrativeArea,
+    this.subAdministrativeArea,
+    this.postalCode,
+    this.country,
+    this.isoCountryCode,
+  });
 
   factory ExpenseLocation.fromJson(Map<String, dynamic> json) {
     return ExpenseLocation(
@@ -12,6 +37,15 @@ class ExpenseLocation {
       longitude: json['longitude']?.toDouble(),
       address: json['address'],
       name: json['name'],
+      street: json['street'],
+      streetNumber: json['streetNumber'],
+      locality: json['locality'],
+      subLocality: json['subLocality'],
+      administrativeArea: json['administrativeArea'],
+      subAdministrativeArea: json['subAdministrativeArea'],
+      postalCode: json['postalCode'],
+      country: json['country'],
+      isoCountryCode: json['isoCountryCode'],
     );
   }
 
@@ -20,6 +54,16 @@ class ExpenseLocation {
     if (longitude != null) 'longitude': longitude,
     if (address != null) 'address': address,
     if (name != null) 'name': name,
+    if (street != null) 'street': street,
+    if (streetNumber != null) 'streetNumber': streetNumber,
+    if (locality != null) 'locality': locality,
+    if (subLocality != null) 'subLocality': subLocality,
+    if (administrativeArea != null) 'administrativeArea': administrativeArea,
+    if (subAdministrativeArea != null)
+      'subAdministrativeArea': subAdministrativeArea,
+    if (postalCode != null) 'postalCode': postalCode,
+    if (country != null) 'country': country,
+    if (isoCountryCode != null) 'isoCountryCode': isoCountryCode,
   };
 
   bool get hasLocation => latitude != null && longitude != null;
@@ -38,12 +82,31 @@ class ExpenseLocation {
     double? longitude,
     String? address,
     String? name,
+    String? street,
+    String? streetNumber,
+    String? locality,
+    String? subLocality,
+    String? administrativeArea,
+    String? subAdministrativeArea,
+    String? postalCode,
+    String? country,
+    String? isoCountryCode,
   }) {
     return ExpenseLocation(
       latitude: latitude ?? this.latitude,
       longitude: longitude ?? this.longitude,
       address: address ?? this.address,
       name: name ?? this.name,
+      street: street ?? this.street,
+      streetNumber: streetNumber ?? this.streetNumber,
+      locality: locality ?? this.locality,
+      subLocality: subLocality ?? this.subLocality,
+      administrativeArea: administrativeArea ?? this.administrativeArea,
+      subAdministrativeArea:
+          subAdministrativeArea ?? this.subAdministrativeArea,
+      postalCode: postalCode ?? this.postalCode,
+      country: country ?? this.country,
+      isoCountryCode: isoCountryCode ?? this.isoCountryCode,
     );
   }
 
@@ -55,9 +118,30 @@ class ExpenseLocation {
           latitude == other.latitude &&
           longitude == other.longitude &&
           address == other.address &&
-          name == other.name;
+          name == other.name &&
+          street == other.street &&
+          streetNumber == other.streetNumber &&
+          locality == other.locality &&
+          subLocality == other.subLocality &&
+          administrativeArea == other.administrativeArea &&
+          subAdministrativeArea == other.subAdministrativeArea &&
+          postalCode == other.postalCode &&
+          country == other.country &&
+          isoCountryCode == other.isoCountryCode;
 
   @override
   int get hashCode =>
-      latitude.hashCode ^ longitude.hashCode ^ address.hashCode ^ name.hashCode;
+      latitude.hashCode ^
+      longitude.hashCode ^
+      address.hashCode ^
+      name.hashCode ^
+      street.hashCode ^
+      streetNumber.hashCode ^
+      locality.hashCode ^
+      subLocality.hashCode ^
+      administrativeArea.hashCode ^
+      subAdministrativeArea.hashCode ^
+      postalCode.hashCode ^
+      country.hashCode ^
+      isoCountryCode.hashCode;
 }